Limiting extra stars to "main" repos
After this change, if you depend on a package that is part of a
larger "meta package", the main repository is only starred if
the package is a *direct* dependency.

For example, if you have symfony/routing in your composer.json,
this will star symfony/symfony also. But if you do not depend
on symfony/routing directly (but it's required by another package),
you *only* star symfony/routing.